pyproject.toml
[project]
name = "stellar-telegram-bot"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "overcat", email = "4catcode@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.3",
    "pika>=1.3.2",
    "pymongo>=4.4.0",
    "pydantic>=1.10.9",
    "stellar-sdk>=8.2.1",
    "loguru>=0.7.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "httpx[socks]>=0.24.1",
    "black>=23.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true
